app/requirements.txt
# requirements.txt

fastapi[all]
uvicorn
pytesseract
pillow
redis
pytest